feat(transfer): add block writer with sync interval
